+++ MegaManLofi/IConsoleSprite.h
@@ -13,7 +13,7 @@ namespace MegaManLofi
 
       virtual short GetWidth() const = 0;
       virtual short GetHeight() const = 0;
-      virtual double GetTotalTraversalSeconds() const = 0;
+      virtual float GetTotalTraversalSeconds() const = 0;
       virtual const ConsoleImage& GetCurrentImage() const = 0;
    };
 }
