+++ MegaManLofiTests/mock_GameClock.h
@@ -13,7 +13,7 @@ public:
    MOCK_METHOD( bool, HasMinimumFrameRate, ( ), ( const, override ) );
    MOCK_METHOD( long long, GetLagFrameCount, ( ), ( const, override ) );
    MOCK_METHOD( long long, GetCurrentFrame, ( ), ( const, override ) );
-   MOCK_METHOD( double, GetFrameSeconds, ( ), ( const, override ) );
+   MOCK_METHOD( float, GetFrameSeconds, ( ), ( const, override ) );
    MOCK_METHOD( void, StartFrame, ( ), ( override ) );
    MOCK_METHOD( void, EndFrame, ( ), ( override ) );
    MOCK_METHOD( long long, GetElapsedNanoseconds, ( ), ( const, override ) );
